style: rename "Result" to "Output" because of ambiguity
